✨  SwitchIcon  ♻️  NavigationMenu
💄  StyledList backgroundColor
