+++ examples/design-system/src/components/Switch/index.ts
@@ -1,0 +1,1 @@
+export { Switch } from './Switch'
