+++ examples/design-system/src/pages/index.tsx
@@ -6,7 +6,8 @@ import { PageHeading, Text } from '@jeromefitz/design-system/src/components'
 // import { LoadingDots } from '../components/LoadingDots'
 import { NavigationMenu } from '../components/NavigationMenu'
 // import { Select } from '../components/Select'
-import { Toast } from '../components/Toast'
+import { Switch } from '../components/Switch'
+// import { Toast } from '../components/Toast'
 // import { ToggleGroup } from '../components/ToggleGroup'
 // import { Toolbar } from '../components/Toolbar'
 // import { LOTS_O_TEXT } from '../lib/constants'
@@ -21,7 +22,8 @@ function Home({}) {
       {/* <LoadingDots /> */}
       <NavigationMenu />
       {/* <Select /> */}
-      <Toast />
+      {/* <Toast /> */}
+      <Switch />
       {/* <ToggleGroup /> */}
       {/* <Toolbar /> */}
       {/* <Text css={{ lineHeight: '1.25' }} size="5">
