+++ packages/design-system/src/components/Switch/index.ts
@@ -1,1 +1,2 @@
 export { Switch } from './Switch'
+export { SwitchIcon } from './SwitchIcon'
