+++ packages/design-system/src/components/index.ts
@@ -109,24 +109,24 @@ export {
   MenuSeparator,
 } from './Menu'
 export {
-  // @core
+  Focused,
   NavigationMenu,
-  NavigationMenuList,
-  NavigationMenuItem,
-  NavigationMenuTrigger,
-  NavigationMenuLink,
+  NavigationMenuCaret,
   NavigationMenuContent,
-  NavigationMenuViewport,
   NavigationMenuIndicator,
-  // @custom
+  NavigationMenuItem,
+  NavigationMenuLink,
+  NavigationMenuLinkText,
+  NavigationMenuLinkTitle,
+  NavigationMenuList,
   NavigationMenuListContent,
   NavigationMenuListItem,
-  NavigationMenuLinkTitle,
-  NavigationMenuLinkText,
-  NavigationMenuViewportPosition,
   NavigationMenuListItemFocus,
   NavigationMenuListItemSelect,
-  Focused,
+  NavigationMenuTrigger,
+  NavigationMenuTriggerWithCaret,
+  NavigationMenuViewport,
+  NavigationMenuViewportPosition,
   Selected,
 } from './NavigationMenu'
 export { Note } from './Note'
@@ -182,7 +182,7 @@ export { Spacer } from './Spacer'
 export { Status } from './Status'
 export { Sub } from './Sub'
 export { Sup } from './Sup'
-export { Switch } from './Switch'
+export { Switch, SwitchIcon } from './Switch'
 export { TableCaption, Tbody, Tfoot, Tr, Th, Td, Thead, Table } from './Table'
 export { TabLink } from './TabLink'
 export { TabsList, Tabs, TabsContent, TabsTrigger } from './Tabs'
